Restructure cache classes for future commit
