+++ src/main/java/net/minecraftforge/eventbus/api/EventListenerHelper.java
@@ -4,22 +4,21 @@
  */
 package net.minecraftforge.eventbus.api;
 
-import net.minecraftforge.eventbus.InternalUtils;
 import net.minecraftforge.eventbus.ListenerList;
 import net.minecraftforge.eventbus.api.Event.HasResult;
+import net.minecraftforge.eventbus.internal.Cache;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Modifier;
-import java.util.function.BiFunction;
-import java.util.function.Supplier;
 
 public class EventListenerHelper {
-    private static final BiFunction<Class<?>, Supplier<ListenerList>, ListenerList> listeners = InternalUtils.cachePublic();
+    private static final Cache<Class<?>, ListenerList> listeners = Cache.create();
     private static final ListenerList EVENTS_LIST = new ListenerList();
-    private static final BiFunction<Class<?>, Supplier<Boolean>, Boolean> cancelable = InternalUtils.cachePublic();
-    private static final BiFunction<Class<?>, Supplier<Boolean>, Boolean> hasResult = InternalUtils.cachePublic();
+    private static final Cache<Class<?>, Boolean> cancelable = Cache.create();
+    private static final Cache<Class<?>, Boolean> hasResult = Cache.create();
+
     /**
      * Returns a {@link ListenerList} object that contains all listeners
      * that are registered to this event class.
@@ -35,7 +34,7 @@ public class EventListenerHelper {
 
     static ListenerList getListenerListInternal(Class<?> eventClass, boolean fromInstanceCall) {
         if (eventClass == Event.class) return EVENTS_LIST; // Small optimization, bypasses all the locks/maps.
-        return listeners.apply(eventClass, () -> computeListenerList(eventClass, fromInstanceCall));
+        return listeners.computeIfAbsent(eventClass, () -> computeListenerList(eventClass, fromInstanceCall));
     }
 
     private static ListenerList computeListenerList(Class<?> eventClass, boolean fromInstanceCall) {
@@ -66,11 +65,11 @@ public class EventListenerHelper {
         return hasAnnotation(eventClass, HasResult.class, hasResult);
     }
 
-    private static boolean hasAnnotation(Class<?> eventClass, Class<? extends Annotation> annotation, BiFunction<Class<?>, Supplier<Boolean>, Boolean> cache) {
+    private static boolean hasAnnotation(Class<?> eventClass, Class<? extends Annotation> annotation, Cache<Class<?>, Boolean> cache) {
         if (eventClass == Event.class || eventClass == Object.class)
             return false;
 
-        return cache.apply(eventClass, () -> {
+        return cache.computeIfAbsent(eventClass, () -> {
             if (eventClass.isAnnotationPresent(annotation))
                 return true;
             var parent = eventClass.getSuperclass();
