+++ src/main/java/net/minecraftforge/eventbus/internal/CacheCopyOnWrite.java
@@ -2,7 +2,7 @@
  * Copyright (c) Forge Development LLC
  * SPDX-License-Identifier: LGPL-2.1-only
  */
-package net.minecraftforge.eventbus;
+package net.minecraftforge.eventbus.internal;
 
 import java.util.Map;
 import java.util.function.Function;
