+++ src/main/java/net/minecraftforge/eventbus/internal/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Internal classes that may change or be removed at any time without notice. Only exported for tests and may no longer
+ * be exported in a future release.
+ */
+@ApiStatus.Internal
+package net.minecraftforge.eventbus.internal;
+
+import org.jetbrains.annotations.ApiStatus;
